fix: unlockable touchups
