+++ components/collection/unlockable/CountdownTimer.vue
@@ -19,12 +19,11 @@
 import 'keen-slider/keen-slider.min.css'
 import { useKeenSlider } from 'keen-slider/vue.es'
 import { countDownTime } from './const'
+import { useCountDown } from './utils/useCountDown'
 const animation = { duration: 50000, easing: (t) => t }
 
-const hours = ref(0)
-const minutes = ref(0)
-const seconds = ref(0)
-const timer = ref()
+const { hours, minutes, seconds } = useCountDown(countDownTime)
+
 const [wrapper] = useKeenSlider({
   loop: true,
   renderMode: 'performance',
@@ -43,24 +42,6 @@ const [wrapper] = useKeenSlider({
   animationEnded(s) {
     s.moveToIdx(s.track.details.abs + 1, true, animation)
   },
-})
-
-const countdown = () => {
-  // Get today's date and time
-  const now = new Date().getTime()
-
-  let distance = countDownTime - now
-  hours.value = Math.floor(distance / (1000 * 60 * 60))
-  minutes.value = Math.floor((distance % (1000 * 60 * 60)) / (1000 * 60))
-  seconds.value = Math.floor((distance % (1000 * 60)) / 1000)
-}
-
-onMounted(() => {
-  timer.value = setInterval(countdown, 1000)
-})
-
-onBeforeMount(() => {
-  clearInterval(timer.value)
 })
 </script>
