+++ components/collection/unlockable/UnlockableCollectionInfo.vue
@@ -1,6 +1,6 @@
 <template>
   <div
-    class="is-flex is-justify-content-space-between mobile-flex-direction-column gap">
+    class="is-flex is-justify-content-space-between mobile-flex-direction-column">
     <div class="is-flex is-flex-direction-column is-flex-grow-1 max-width">
       <div class="is-flex mt-6 is-justify-content-space-between mb-2">
         <div class="mr-2 has-text-weight-bold is-size-5 mb-1">
@@ -19,23 +19,24 @@
           @click.native="toggleSeeAllDescription" />
         <div v-else />
 
-        <nuxt-link
+        <NeoButton
+          variant="unlockable"
+          tag="nuxt-link"
           :to="`/${urlPrefix}/collection/${collectionId}`"
-          target="_blank">
-          <NeoButton variant="unlockable" class="has-text-left mt-4">
-            View Collection
-            <svg
-              width="12"
-              height="12"
-              viewBox="0 0 12 12"
-              fill="none"
-              xmlns="http://www.w3.org/2000/svg">
-              <path
-                d="M1.41708 6.95833L0.95875 6.95833L0.95875 6.04167L1.41708 6.04167L1.41708 6.95833ZM10.9078 6.1759C11.0868 6.35489 11.0868 6.64509 10.9078 6.82408L7.99103 9.7409C7.81204 9.91989 7.52183 9.91989 7.34284 9.7409C7.16385 9.56191 7.16385 9.27171 7.34284 9.09272L9.93557 6.49999L7.34284 3.90727C7.16385 3.72828 7.16385 3.43808 7.34284 3.25909C7.52183 3.0801 7.81204 3.0801 7.99103 3.25909L10.9078 6.1759ZM1.41708 6.04167L10.5838 6.04166L10.5838 6.95833L1.41708 6.95833L1.41708 6.04167Z"
-                fill="currentColor" />
-            </svg>
-          </NeoButton>
-        </nuxt-link>
+          target="_blank"
+          class="has-text-left">
+          View Collection
+          <svg
+            width="12"
+            height="12"
+            viewBox="0 0 12 12"
+            fill="none"
+            xmlns="http://www.w3.org/2000/svg">
+            <path
+              d="M1.41708 6.95833L0.95875 6.95833L0.95875 6.04167L1.41708 6.04167L1.41708 6.95833ZM10.9078 6.1759C11.0868 6.35489 11.0868 6.64509 10.9078 6.82408L7.99103 9.7409C7.81204 9.91989 7.52183 9.91989 7.34284 9.7409C7.16385 9.56191 7.16385 9.27171 7.34284 9.09272L9.93557 6.49999L7.34284 3.90727C7.16385 3.72828 7.16385 3.43808 7.34284 3.25909C7.52183 3.0801 7.81204 3.0801 7.99103 3.25909L10.9078 6.1759ZM1.41708 6.04167L10.5838 6.04166L10.5838 6.95833L1.41708 6.95833L1.41708 6.04167Z"
+              fill="currentColor" />
+          </svg>
+        </NeoButton>
       </div>
     </div>
     <div>
