+++ libs/ui/src/components/NeoButton/NeoButton.scss
@@ -105,7 +105,7 @@
   }
 
   &--unlockable{
-    height: 40px;
+    height: 30px;
     width: min-content !important;
     border-radius: 4rem;
     box-shadow: none !important;
